changed value type of Forecast "offset" to double from int

+++ app/src/main/java/com/shawnaten/networking/Forecast.java
@@ -23,7 +23,7 @@ public class Forecast {
 
         private double latitude, longitude;
         private TimeZone timezone;
-        private int offset;
+        private double offset;
 
         private DataPoint currently;
 
@@ -62,7 +62,7 @@ public class Forecast {
             return timezone;
         }
 
-        public int getOffset() {
+        public double getOffset() {
             return offset;
         }
